Move port functions to C file
